Add files via upload
This repository provides code for the analyses presented in the manuscript submitted to PNAS.  

The main m.file executes simulations on the stimulation of a neuron. The execution requires a definition of the so-called control signal–induced current in a cell. Here, we consider a current imported from the enclosed .xlsx file.